ÚLTIMO CAMBIO DEL DÍA

+++ Presentación-GUI/FrmGestionarEmbalaje.cs
@@ -51,9 +51,12 @@ namespace Presentación_GUI
             vlr5 = decimal.Parse(TxtAnchoEmbalaje.Text);
             vlr6 = decimal.Parse(TxtAltoEmbalaje.Text);
 
-            total1 = vlr4/vlr1;
-            total2 = vlr5/vlr2;
-            total3 = vlr6/vlr3;
+            
+
+            total1 = (int)vlr4/ (int)vlr1;
+            total2 = (int)vlr5 / (int)vlr2;
+            total3 = (int)vlr6 / (int)vlr3;
+
             
             totaltotal = total1 * total2 * total3;
